refactor commands by categories
