+++ src/commands/files/add.js
