+++ src/commands/files/cat.js
