+++ src/commands/files/cp.js
