+++ src/commands/files/rm.js
